chore: separate external data tests
